Perform scPAS Screening Analysis

This function performs scPAS screening analysis by integrating bulk and single-cell RNA-seq data. It includes data filtering steps and wraps the core scPAS::scPAS function.

Usage

DoscPAS(
  matched_bulk,
  sc_data,
  phenotype,
  label_type = "scPAS",
  assay = "RNA",
  imputation = F,
  imputation_method = c("KNN", "ALRA"),
  nfeature = 3000,
  alpha = 0.01,
  network_class = c("SC", "bulk"),
  scPAS_family = c("cox", "gaussian", "binomial"),
  permutation_times = 2000,
  FDR.threshold = 0.05,
  independent = TRUE,
  ...
)

Arguments

matched_bulk

Bulk RNA-seq data (genes x samples)

sc_data

Single-cell RNA-seq data (Seurat object and preprocessed)

phenotype

Phenotype data frame with sample annotations

label_type

Character specifying phenotype label type (e.g., "SBS1", "time"), stored in scRNA_data@misc

assay

Assay to use from sc_data (default: 'RNA')

imputation

Logical, whether to perform imputation (default: FALSE)

imputation_method

Character. Name of alternative method for imputation.

nfeature

Number of features to select (default: 3000, indicating that the top 3000 highly variable genes are selected for model training

alpha

Numeric. Significance threshold. Parameter used to balance the effect of the l1 norm and the network-based penalties. It can be a number or a searching vector. If alpha = NULL, a default searching vector is used. The range of alpha is in [0,1]. A larger alpha lays more emphasis on the l1 norm. (default: 0.01)

network_class

Network class to use (default: 'SC', indicating gene-gene similarity networks derived from single-cell data. The other one is 'bulk'.)

scPAS_family

Model family for analysis (options: "cox", "gaussian", "binomial")

permutation_times

Number of permutations to perform (default: 2000)

FDR.threshold

Numeric. FDR value threshold for identifying phenotype-associated cells (default: 0.05)

...

Additional arguments passed to DoscPAS functions

Value

A Seurat object from scPAS analysis

See also