Optimized Scissor Algorithm for Seurat ver5

Scissor.v5 from https://doi.org/10.1038/s41587-021-01091-3and https://github.com/sunduanchen/Scissor/issues/59 Another version of Scissor.v5() to optimize memory usage and execution speed in preprocess.

Usage

Scissor.v5.optimized(
  bulk_dataset,
  sc_dataset,
  phenotype,
  tag = NULL,
  alpha = NULL,
  cutoff = 0.2,
  family = c("gaussian", "binomial", "cox"),
  Save_file = "Scissor_inputs.RData",
  Load_file = NULL,
  ...
)

LICENSE

Licensed under the GNU General Public License version 3 (GPL-3.0). A copy of the license is available at https://www.gnu.org/licenses/gpl-3.0.en.html.

References

Sun D, Guan X, Moran AE, Wu LY, Qian DZ, Schedin P, et al. Identifying phenotype-associated subpopulations by integrating bulk and single-cell sequencing data. Nat Biotechnol. 2022 Apr;40(4):527–38.

See also