Registry of Phenotype-Associated Cell Screening Methods

An environment storing methods for screening phenotype-associated cells.

Usage

ScreenStrategy

Format

An object of class environment of length 7.

Details

Storing structure - named list

key: method name

value: list

method_name: method name

executor: function implementation of the method

phenotypes: The phenotype types supported by this method

mapper: A function that transforms the parameters passed to the Screen function before forwarding them to the executor; both input and output must be of type list.